Refactor the ugly code of the Player implementation to another class
